move ConfiguredEngine, ConfiguredClusterNode to engine.py
